# auth/auth-ws/src/main/resources/openapi/cp-users-api.yml
openapi: 3.0.3
info:
  title: Insureth Client Portal Users API
  description: Authentication and Client Portal User management for Insureth
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Local development server
paths:
  /cp-users/exists:
    get:
      summary: Check if a client user profile exists
      description: Check if a client user profile exists
      operationId: checkClientUserExists
      parameters:
        - name: walletAddress
          in: query
          required: true
          description: The 42-character Ethereum wallet address (e.g., beginning with 0x).
          schema:
            type: string
            example: "0x71C7656EC7ab88b098defB751B7401B5f6d8976F"
      responses:
        "200":
          description: Returns true if username exists
          content:
            application/json:
              schema:
                type: boolean
  /cp-users/create:
    post:
      summary: Creates a new client user
      description: Creates a new client user
      operationId: createClientUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientUserModel'
      responses:
        "201":
          description: User created successfully

  /cp-users/{walletAddress}:
    get:
      summary: Get a client user profile
      description: Retrieves the user's profile details.
      operationId: getClientUser
      parameters:
        - name: walletAddress
          in: path
          required: true
          description: The 42-character Ethereum wallet address.
          schema:
            type: string
            example: "0x71C7656EC7ab88b098defB751B7401B5f6d8976F"
      responses:
        "200":
          description: Retuns the user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientUserModel'
        "404":
          description: User not found
    put:
      summary: Update a client user profile
      description: Updates the user's username and email.
      operationId: updateClientUser
      parameters:
        - name: walletAddress
          in: path
          required: true
          description: The 42-character Ethereum wallet address.
          schema:
            type: string
            example: "0x71C7656EC7ab88b098defB751B7401B5f6d8976F"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientUserModel'
      responses:
        "200":
          description: User updated successfully
        "404":
          description: User not found

components:
  schemas:
    ClientUserModel:
      type: object
      properties:
        walletAddress:
          type: string
          example: "0x71C7656EC7ab88b098defB751B7401B5f6d8976F"
        username:
          type: string
          example: "user01"
        email:
          type: string
          example: "email@gmail.com"